+++ packages/sdk/src/pub/plt.ts
@@ -1,6 +1,8 @@
 import * as Token from '../plt/Token.js';
 // To limit the exports meant only for internal use, we re-create the module exports.
 import * as Cbor from './plt/Cbor.js';
+import * as CborAccountAddress from './plt/CborAccountAddress.js';
+import * as CborContractAddress from './plt/CborContractAddress.js';
 import * as CborMemo from './plt/CborMemo.js';
 import * as TokenAmount from './plt/TokenAmount.js';
 import * as TokenHolder from './plt/TokenHolder.js';
@@ -14,4 +16,15 @@ export * from '../plt/TokenModuleRejectReason.js';
 export * from '../plt/TokenModuleEvent.js';
 export * from '../plt/TokenOperation.js';
 
-export { Token, Cbor, TokenAmount, CborMemo, TokenId, TokenModuleReference, TokenMetadataUrl, TokenHolder };
+export {
+    Token,
+    Cbor,
+    TokenAmount,
+    CborMemo,
+    TokenId,
+    TokenModuleReference,
+    TokenMetadataUrl,
+    TokenHolder,
+    CborAccountAddress,
+    CborContractAddress,
+};