+++ packages/sdk/src/types/transactionEvent.ts
@@ -1,5 +1,5 @@
-import { Upward } from '../grpc/index.js';
-import * as PLT from '../plt/index.js';
+import type { Upward } from '../grpc/index.js';
+import type * as PLT from '../plt/index.js';
 import type {
     Address,
     BakerId,
@@ -384,10 +384,20 @@ export type TokenTransferEvent = {
     tag: TransactionEventTag.TokenTransfer;
     /** The token ID of the token the event originates from */
     tokenId: PLT.TokenId.Type;
-    /** The token holder sending the tokens. */
-    from: PLT.TokenHolder.Type;
-    /** The token holder receiving the tokens. */
-    to: PLT.TokenHolder.Type;
+    /**
+     * The token holder sending the tokens.
+     *
+     * **Please note**, this can possibly be unknown if the SDK is not fully compatible with the Concordium
+     * node queried, in which case `null` is returned.
+     */
+    from: Upward<PLT.TokenHolder.Type>;
+    /**
+     * The token holder receiving the tokens.
+     *
+     * **Please note**, this can possibly be unknown if the SDK is not fully compatible with the Concordium
+     * node queried, in which case `null` is returned.
+     */
+    to: Upward<PLT.TokenHolder.Type>;
     /** The amount of tokens transferred. */
     amount: PLT.TokenAmount.Type;
     /** An optional memo associated with the transfer. */
@@ -402,8 +412,13 @@ export type TokenMintEvent = {
     tag: TransactionEventTag.TokenMint;
     /** The token ID of the token the event originates from */
     tokenId: PLT.TokenId.Type;
-    /** The token holder whose supply is updated. */
-    target: PLT.TokenHolder.Type;
+    /**
+     * The token holder whose supply is updated.
+     *
+     * **Please note**, this can possibly be unknown if the SDK is not fully compatible with the Concordium
+     * node queried, in which case `null` is returned.
+     */
+    target: Upward<PLT.TokenHolder.Type>;
     /** The amount by which the token supply is updated. */
     amount: PLT.TokenAmount.Type;
 };
@@ -416,8 +431,13 @@ export type TokenBurnEvent = {
     tag: TransactionEventTag.TokenBurn;
     /** The token ID of the token the event originates from */
     tokenId: PLT.TokenId.Type;
-    /** The token holder whose supply is updated. */
-    target: PLT.TokenHolder.Type;
+    /**
+     * The token holder whose supply is updated.
+     *
+     * **Please note**, this can possibly be unknown if the SDK is not fully compatible with the Concordium
+     * node queried, in which case `null` is returned.
+     */
+    target: Upward<PLT.TokenHolder.Type>;
     /** The amount by which the token supply is updated. */
     amount: PLT.TokenAmount.Type;
 };